+++ Tools/PC/jira-card-creator/.gitignore
@@ -1,6 +1,6 @@
 jira-creator-env
 Jira
-Google
+GoogleSheet
 credentials.json
 token.json
 service-account-key.json
